+++ CHANGELOG.md
@@ -5,6 +5,10 @@
 - Allow cancelling screenshots
 - `move_dist` command
 - Value control commands without args now print out the current value
+- `zoom_lvl` command
+
+# Fixed
+- Some errors in help messages
 
 # 0.3.1 - 2023-11-06